+++ replace_test.go
@@ -37,13 +37,13 @@ func TestReplace(t *testing.T) {
 
 	data, err := conn.Execute(&Replace{
 		Space: "tester",
-		Tuple: []interface{}{4, "Hello"},
+		Tuple: []interface{}{uint64(4), "Hello"},
 	})
 
 	if assert.NoError(err) {
 		assert.Equal([]interface{}{
 			[]interface{}{
-				uint32(4),
+				uint64(4),
 				"Hello",
 			},
 		}, data)
@@ -51,13 +51,13 @@ func TestReplace(t *testing.T) {
 
 	data, err = conn.Execute(&Replace{
 		Space: "tester",
-		Tuple: []interface{}{4, "World"},
+		Tuple: []interface{}{uint64(4), "World"},
 	})
 
 	if assert.NoError(err) {
 		assert.Equal([]interface{}{
 			[]interface{}{
-				uint32(4),
+				uint64(4),
 				"World",
 			},
 		}, data)
